re-seeded with realistic data migration

+++ Domain/ProductClass.cs
@@ -11,12 +11,12 @@ namespace Domain
         public Boolean onSale { get; set; }
         public string soldBy { get; set; }
         public int caseSize { get; set; }
-        public int retailPrice { get; set; }
-        public int sold { get; set; }
-        public int shrink { get; set; }
-        public int expectedInv { get; set; }
-        public int expectedFloor { get; set; }
-        public int par { get; set; }
+        public double retailPrice { get; set; }
+        public double sold { get; set; }
+        public double shrink { get; set; }
+        public double expectedInv { get; set; }
+        public double expectedFloor { get; set; }
+        public double par { get; set; }
         public Boolean productActive { get; set; }
     }
 }
